feat: Add hierarchy_level field to Role model
This commit adds a new field, `hierarchy_level`, to the `Role` model in the `about/models.py` file. The `hierarchy_level` field is an integer field with a default value of 100. It is used to specify the hierarchy level of roles, which will be used for ordering them in the admin interface and other views.