feat: Add views_count field to BaseContent model
The commit adds a new field `views_count` to the `BaseContent` model in the `npf/blog/models.py` file. This field is a positive integer field with a default value of 0. It will be used to track the number of views for each content item.

Note: Please remove any additional explanations or details from the commit message.